test(signing): split signer tests per method and add signer unit tests
